feat: implementa Practitioner refactor en service, repository y controller
